修改了一处导致警告的地方

+++ LR(1)_Creater/main.cpp
@@ -17,7 +17,7 @@ int main()
 		G.Push_Back(s);
 		Grammar_List.push_back(s);
 	}
-	fprintf(fp,"%d\n",Grammar_List.size());
+	fprintf(fp,"%d\n",(int)Grammar_List.size());
 	for(int i=0;i<Grammar_List.size();i++)
 		fprintf(fp,"%s\n",Grammar_List[i].c_str());
